moved plotly stuff to its own plugin
