+++ tidy3d/plugins/plotly/__init__.py
@@ -1,0 +1,2 @@
+from .geo import SimulationPlotly
+from .data import *
